+++ lib/pymedphys/_dicom/dose.py
@@ -14,12 +14,16 @@
 
 """A DICOM RT Dose toolbox"""
 
+import copy
+from typing import Sequence
+
 from pymedphys._imports import matplotlib
 from pymedphys._imports import numpy as np
 from pymedphys._imports import plt, pydicom, scipy
 
 from .constants import IMAGE_ORIENTATION_MAP
-from .coords import xyz_axes_from_dataset
+from .coords import coords_in_datasets_are_equal, xyz_axes_from_dataset
+from .header import patient_ids_in_datasets_are_equal
 from .rtplan import get_surface_entry_point_with_fallback, require_gantries_be_zero
 from .structure import pull_structure
 
@@ -314,3 +318,73 @@ def require_patient_orientation(ds, patient_orientation):
             "The supplied dataset has a patient "
             f"orientation other than {patient_orientation}."
         )
+
+
+def sum_doses_in_datasets(
+    datasets: Sequence["pydicom.dataset.Dataset"],
+) -> "pydicom.dataset.Dataset":
+    """Sum two or more DICOM dose grids and save to new DICOM RT
+    Dose dataset"
+
+    Parameters
+    ----------
+    datasets : sequence of pydicom.dataset.Dataset
+        A sequence of DICOM RT Dose datasets whose doses are to be
+        summed.
+
+    Returns
+    -------
+    pydicom.dataset.Dataset
+        A new DICOM RT Dose dataset whose dose is the sum of all doses
+        within `datasets`
+    """
+
+    if not len(datasets) >= 2:
+        raise ValueError("At least two datasets must be provided for comparison")
+
+    if not all(ds.Modality == "RTDOSE" for ds in datasets):
+        raise ValueError("`datasets` must only contain DICOM RT Dose datasets.")
+
+    if not patient_ids_in_datasets_are_equal(datasets):
+        raise ValueError("Patient ID must match for all datasets")
+
+    if not all(ds.DoseSummationType == "PLAN" for ds in datasets):
+        raise ValueError(
+            "Only DICOM RT Doses whose DoseSummationTypes are 'PLAN' are supported"
+        )
+
+    if not all(ds.DoseUnits == datasets[0].DoseUnits for ds in datasets):
+        raise ValueError(
+            "All DICOM RT Doses must have the same units ('GY or 'RELATIVE')"
+        )
+
+    if not coords_in_datasets_are_equal(datasets):
+        raise ValueError("All dose grids must have perfectly coincident coordinates")
+
+    ds_summed = copy.deepcopy(datasets[0])
+
+    ds_summed.BitsAllocated = 32
+    ds_summed.BitsStored = 32
+    ds_summed.DoseSummationType = "MULTI_PLAN"
+    ds_summed.DoseComment = "Summed Dose"
+
+    if not all(ds.DoseType in ("PHYSICAL", "EFFECTIVE") for ds in datasets):
+        raise ValueError(
+            "Only DICOM RT Doses whose DoseTypes are 'PHYSICAL' or "
+            "'EFFECTIVE' are supported"
+        )
+    if any(ds.DoseType == "EFFECTIVE" for ds in datasets):
+        ds_summed.DoseType = "EFFECTIVE"
+    else:
+        ds_summed.DoseType = "PHYSICAL"
+    doses = np.array([dose_from_dataset(ds) for ds in datasets])
+    doses_summed = np.sum(doses, axis=0, dtype=np.float32)
+
+    # ds_summed.DoseGridScaling = np.max(doses_summed) / (2 ^ int(ds_summed.HighBit))
+
+    ds_summed.DoseGridScaling = np.max(doses_summed) / (2 ** 32 - 1)
+    pixel_array_summed = (doses_summed / ds_summed.DoseGridScaling).astype(np.uint32)
+
+    ds_summed.PixelData = pixel_array_summed.tobytes()
+
+    return ds_summed
