+++ lib/pymedphys/_dicom/header.py
@@ -15,6 +15,7 @@
 
 import re
 from copy import deepcopy
+from typing import Sequence
 
 from pymedphys._imports import pydicom
 
@@ -217,3 +218,26 @@ def pretty_patient_name(
         pretty_name = f"{honorific}. {pretty_name}"
 
     return pretty_name
+
+
+def patient_ids_in_datasets_are_equal(
+    datasets: Sequence["pydicom.dataset.Dataset"],
+) -> bool:
+    """True if all DICOM datasets have the same Patient ID
+
+    Parameters
+    ----------
+    datasets : sequence of pydicom.dataset.Dataset
+        A sequence of DICOM datasets whose Patient IDs are to be
+        compared.
+
+    Returns
+    -------
+    bool
+        True if Patient IDs match for all datasets, False otherwise.
+    """
+
+    if not len(datasets) >= 2:
+        raise ValueError("At least two datasets must be provided for comparison")
+
+    return all(ds.PatientID == datasets[0].PatientID for ds in datasets)
